helix_backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
firebase-admin>=6.4.0
django-cors-headers>=4.3.0
django-filter>=23.5
psycopg2-binary>=2.9.9
python-decouple>=3.8
